requirements.txt
tensorflow
keras
yfinance
pandas
numpy
scikit-learn
mysql-connector-python